+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Emit event when body contact is added
 - Add `:vf/unique` component trait, it forces a component to be applied to one entity
   only, removing it from all the others entities
-- Add OnRaycastClick and OnRaycastHover
+- Add on click and on hover events
 
 ## v0.1.75